# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Journey API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/journey/{journey}/activity/{activity}/action": {
            "get": {
                "tags": [
                    "journeyAction"
                ],
                "summary": "List actions",
                "description": "Get a list of all actions of a journey activity.",
                "operationId": "journeyActionList",
                "parameters": [
                    {
                        "name": "journey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "activity",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/JourneyAction"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The journey or activity does not exist."
                    }
                }
            },
            "post": {
                "tags": [
                    "journeyAction"
                ],
                "summary": "Create activity",
                "description": "Create a new action for the journey activity.",
                "operationId": "journeyActionCreate",
                "parameters": [
                    {
                        "name": "journey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "activity",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "action_id",
                        "in": "query",
                        "description": "The reference to the action the user starts.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "started_at",
                        "in": "query",
                        "description": "The date and time the action was started.",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "status_id",
                        "in": "query",
                        "description": "The reference to the status of the action.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "input_id",
                        "in": "query",
                        "description": "The reference to the input attached to the action.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JourneyAction"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The entered parameters are not valid."
                    },
                    "404": {
                        "description": "The journey or activity does not exist."
                    }
                }
            }
        },
        "/journey/{journey}/activity/{activity}/action/{id}": {
            "get": {
                "tags": [
                    "journeyAction"
                ],
                "summary": "Show action",
                "description": "Get a single action from a journey activity by id.",
                "operationId": "journeyActionShow",
                "parameters": [
                    {
                        "name": "journey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "activity",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JourneyAction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The journey, activity or action does not exist."
                    }
                }
            },
            "delete": {
                "tags": [
                    "journeyAction"
                ],
                "summary": "Delete action",
                "description": "Delete an action of a journey activity.",
                "operationId": "journeyActionDelete",
                "parameters": [
                    {
                        "name": "journey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "activity",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "The journey, activity or action does not exist."
                    }
                }
            },
            "patch": {
                "tags": [
                    "journeyAction"
                ],
                "summary": "Update action",
                "description": "Update an existing action of a journey activity.",
                "operationId": "journeyActionUpdate",
                "parameters": [
                    {
                        "name": "journey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "activity",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "action_id",
                        "in": "query",
                        "description": "The reference to the action the user starts.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "started_at",
                        "in": "query",
                        "description": "The date and time the action was started.",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "status_id",
                        "in": "query",
                        "description": "The reference to the status of the action.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "input_id",
                        "in": "query",
                        "description": "The reference to the input attached to the action.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JourneyAction"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The entered parameters are not valid."
                    },
                    "404": {
                        "description": "The journey, activity or action does not exist."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Action": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "activity": {
                        "$ref": "#/components/schemas/Activity"
                    },
                    "input_type": {
                        "$ref": "#/components/schemas/InputType"
                    },
                    "input_required": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "name": {
                        "type": "string",
                        "example": "Take Photo"
                    },
                    "description": {
                        "type": "string",
                        "example": "Now the photo will be taken by the selfie box."
                    },
                    "created_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    },
                    "updated_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Action"
                }
            },
            "Activity": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Photos"
                    },
                    "description": {
                        "type": "string",
                        "example": "Take two photos in the selfie box."
                    },
                    "created_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    },
                    "updated_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Activity"
                }
            },
...
            "Input": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Take Photo"
                    },
                    "data_type_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "data_type_type": {
                        "type": "string",
                        "example": "App\\\\Models\\\\ImageInput"
                    },
                    "created_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    },
                    "updated_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Input"
                }
            },
            "InputType": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Image"
                    },
                    "description": {
                        "type": "string",
                        "example": "The user has to submit an image input."
                    },
                    "class_name": {
                        "type": "string",
                        "example": "App\\\\Models\\\\ImageInput"
                    },
                    "created_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    },
                    "updated_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "InputType"
                }
            },
            "Journey": {
                "required": [
                    "name",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Selfie-box Journey"
                    },
                    "started_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    },
                    "status": {
                        "$ref": "#/components/schemas/Status"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    },
                    "updated_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Journey"
                }
            },
            "JourneyAction": {
                "required": [
                    "journey_activity",
                    "action",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "journey_activity": {
                        "$ref": "#/components/schemas/JourneyActivity"
                    },
                    "action": {
                        "$ref": "#/components/schemas/Action"
                    },
                    "started_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    },
                    "status": {
                        "$ref": "#/components/schemas/Status"
                    },
                    "input": {
                        "$ref": "#/components/schemas/Input"
                    },
                    "created_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    },
                    "updated_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "JourneyAction"
                }
            },
            "JourneyActivity": {
                "required": [
                    "journey",
                    "activity",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "journey": {
                        "$ref": "#/components/schemas/Journey"
                    },
                    "activity": {
                        "$ref": "#/components/schemas/Activity"
                    },
                    "started_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    },
                    "status": {
                        "$ref": "#/components/schemas/Status"
                    },
                    "created_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    },
                    "updated_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "JourneyActivity"
                }
            },
            "Status": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Created"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Status"
                }
            },
...
            "User": {
                "required": [
                    "first_name",
                    "family_name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Jeremy"
                    },
                    "family_name": {
                        "type": "string",
                        "example": "Becker"
                    },
                    "email": {
                        "type": "string",
                        "example": "jeremy@becker.dev"
                    },
                    "password": {
                        "type": "string"
                    },
                    "language": {
                        "type": "string",
                        "example": "de"
                    },
                    "is_admin": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "token": {
                        "type": "string",
                        "example": "d8fah"
                    },
                    "created_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    },
                    "updated_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}